remove ts stuff
