--- tests/formatter.test.ts
@@ -1,4 +1,0 @@
-import { tokenValueToNumber } from "../src/rateApi";
-
-console.log(await tokenValueToNumber("ethereum"));
-console.log(await tokenValueToNumber("usd-coin"));
