--- tests/twitter.test.ts
@@ -1,3 +1,0 @@
-import { tweet } from "../src/twitter";
-
-tweet("SomeText");
